Create Marks of 5 student assignment.
Write a C program to read name and marks of 5 students and store them in a file. 
Write a C program to read name and marks of n number of students from and store them in a file. If the file previously exits, add the information to the file. 
Write a C program to write all the members of an array of structures to a file using fwrite(). Read the array from the file and display on the screen.