Create Library overdue charges assignment.
a) You are required to develop a C program to calculate the fine for overdue library books. The
library charges fines based on the number of days a book is overdue. The fine rates are as
follows:
i. Write a C program that takes the following inputs from the user:Book ID (an integer).
Due Date (an integer)and Return Date (an integer).
(3 marks)
Calculate the days overdue Hint: returnDate-dueDate
(2 marks)
iii. Use an if…else statement to determine the fine rate as per the specified rates in the above
table and apply the necessary calculations.
(6 marks)
iv. The program should display the following
(4 marks)
bookID
dueDate
returnDate
daysOverdue
fineRate
fineAmount