Bank loans to customers assignment.
Read about the conditional statements(if, if…else, else…if)then answer the questions below. 
a) A bank will offer a customer a loan if they are 21 years or over and have an annual income of
at least 
 The customers age and income are input in response to user friendly
prompts. Create a C program to implement the above requirements. (6 marks)
Expected output
i. Congratulations you qualify for a loan.
ii. Unfortunately, we are unable to offer you a loan at this time .